initial attempt at a Windows build
I have found it difficult to get the Windows build working using CMake.
There are several 'best-practices' approaches for getting a CMake project
to link to SDL2 properly, but they would require making some bigger
changes to the project than I feel comfortable with.

Option 1: Use the standard CMake tools to find and link SDL - this would be
"find_package(SDL2 REQUIRED)", and a hit at the local SDL2 winlib path.
This is simple, but it requires a big change:
sed 's/#include <SDL2/SDL.h>/#include <SDL.h>/g'
This is the so-called 'modern' way to include SDL.

Option 2: Require the user to install the SDL2 dev package separately.
This defeats the purpose of including the SDL libs in the first place.

Option 3: Use the Conan package manager. This makes everything dead simple.
You can include any version of lua or stb this way as well, and even keep
those dependencies up to date.

Option 4: say that CMake for this project is a convenience for Linux users
only.